// src/app/models/spell.ts
export interface Spell {
  index: string;
  name: string;
  desc: string[];
  range: string;
  components: string[];
  ritual: boolean;
  duration: string;
  concentration: boolean;
  casting_time: string;
  level: number;
  damage: any;
  dc: any;
  school: any;
  classes: any[];
  subclasses: any[];
  url: string;

}